Refactor nif code to simplify its interface
Using the generic EVP_* API makes it possible to unify
algorithm-specific nif functions to a single generic
function. Effectively the same change that took place on the Erlang
API in R16B01 is now applied to the C code.

As the EVP_* API works both in FIPS and in non-FIPS mode, the separate
non-FIPS code path can be eliminated as long as the public API is not
affected. It had to be kept back for two reasons in some places:

- first, parts of the EVP API were introduced in OpenSSL 1.0.0, in
  which case non-FIPS builds shall still use the low-level API path
- second, the streaming API, which has to use the same context data
  (the low-level API's context structures) to remain backward
  compatible (although this could be removed at the next major
  release)

There are various minor improvements as well:
- supported algorithms are now provided by the nif code (not a mix of
  the C and Erlang code)
- remove unnecessary variables and macro definitions

Note: the disabled algorithm tests in FIPS mode had to be slightly
updated because the new, generic code checks key, iv and data sizes
first.